gdrive-sync/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pypdf>=3.17.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pinecone>=5.0.0
langchain-text-splitters>=0.2.0
schedule>=1.2.0
